menambahkan method delete dan put pada kategori pengaduan

+++ controllers/KatPengaduanControl.js
@@ -4,6 +4,7 @@ import imgParser from "../func/imgParser.js"
 import KategoriPengaduanVal from "../validation/KategoriPengaduanVal.js"
 import ImgVal from '../validation/ImgVal.js'
 import moveUploadedFile from "../func/moveUploadedFile.js"
+import getDirName from "../func/getDirName.js"
 
 class KatPengaduanControl{
     static async getAll(req,res){
@@ -51,7 +52,9 @@ class KatPengaduanControl{
 
             if( typeof parse.files.foto === "undefined"){
                 const keyUp = Object.keys(parse.files)[0]
-                fs.unlinkSync(parse.files[keyUp].filepath)
+                if(keyUp){
+                    fs.unlinkSync(parse.files[keyUp].filepath)
+                }
                 return res.status(400).json({
                     status : "Bad Request",
                     message : "terjadi kesalahan diclient",
@@ -112,6 +115,198 @@ class KatPengaduanControl{
             return res.status(201).json({
                 status : "Created",
                 message : "berhasil menambahkan kategori pengaduan",
+                errors : [],
+                data : []
+            })
+
+
+        }catch(err){
+            return res.status(500).json({
+                status : "Internal Server Error",
+                message : "terjadi kesalahan diserver",
+                errors : [err.message],
+                data : []
+            })
+        }
+    }
+
+    static async put(req,res){
+        try{
+
+            const checkContentType = req.is('multipart/form-data')
+            if(!checkContentType){
+                return res.status(400).json({
+                    status : "Bad Request",
+                    message : "terjadi kesalahan diclient",
+                    errors : ["content type harus multipart/form-data"],
+                    data : []
+                })
+            }
+
+
+            const data = await imgParser(req)
+            const ubah = JSON.stringify(data)
+            const parse = JSON.parse(ubah)
+
+            const {result} = await mysqlQuery(`SELECT * FROM kategori_pengaduan WHERE id = ${req.params.id}`)
+            if(!result.length){
+                const keyUp = Object.keys(parse.files)[0]
+                if(keyUp){
+                    fs.unlinkSync(parse.files[keyUp].filepath)
+                }
+                return res.status(404).json({
+                    status : "Not Found",
+                    message : "terjadi kesalahan diclient",
+                    errors : ["kategori pengaduan tidak ditemukan"],
+                    data : []
+                })
+            }
+
+            if( typeof parse.files.foto === "undefined"){
+
+                const val = new KategoriPengaduanVal(parse.field)
+                val.checkType()
+    
+                if(val.getErrors().length){
+                    return res.status(400).json({
+                        status : "Bad Request",
+                        message : "terjadi kesalahan diclient",
+                        errors : val.getErrors(),
+                        data : []
+                    })
+                }
+                
+                val.checkLen()
+                if(val.getErrors().length){
+                    return res.status(400).json({
+                        status : "Bad Request",
+                        message : "terjadi kesalahan diclient",
+                        errors : val.getErrors(),
+                        data : []
+                    })
+                }
+    
+
+                const sql = `UPDATE kategori_pengaduan SET nama = '${val.nama}', deskripsi = '${val.deskripsi}' WHERE id = ${req.params.id}`
+
+                await mysqlQuery(sql)
+
+                return res.status(200).json({
+                    status : "OK",
+                    message : "berhasil mengubah kategori pengaduan",
+                    errors : [],
+                    data : []
+                })
+
+            }
+
+           
+
+            const urlFileUpload = parse.files.foto.filepath
+            const val = new KategoriPengaduanVal(parse.field)
+            val.checkType()
+
+            if(val.getErrors().length){
+                fs.unlinkSync(urlFileUpload)
+                return res.status(400).json({
+                    status : "Bad Request",
+                    message : "terjadi kesalahan diclient",
+                    errors : val.getErrors(),
+                    data : []
+                })
+            }
+            
+            val.checkLen()
+            
+            if(val.getErrors().length){
+                fs.unlinkSync(urlFileUpload)
+                return res.status(400).json({
+                    status : "Bad Request",
+                    message : "terjadi kesalahan diclient",
+                    errors : val.getErrors(),
+                    data : []
+                })
+            }
+
+
+            
+        
+
+
+            const checkImg = new ImgVal(parse.files.foto)
+            checkImg.checkSize()
+            checkImg.checkIsImg()
+
+            if(checkImg.getErrors().length){
+                fs.unlinkSync(urlFileUpload)
+                return res.status(400).json({
+                    status : "Bad Request",
+                    message : "terjadi kesalahan diclient",
+                    errors : checkImg.getErrors(),
+                    data : []
+                })
+            }
+
+
+            const gambar = moveUploadedFile(parse.files.foto)
+            const imgUrl = `${req.protocol}://${req.headers.host}/gambar/${gambar}`
+        
+            
+            const sql = `UPDATE kategori_pengaduan SET nama = '${val.nama}', foto = '${imgUrl}', deskripsi = '${val.deskripsi}' WHERE id = ${req.params.id}`
+            
+            const dirName = getDirName()
+
+            const imgOld = result[0].foto.split("/")
+            const imgUrlOld = imgOld[imgOld.length - 1]
+
+            fs.unlinkSync(`${dirName}/images/${imgUrlOld}`)
+
+            await mysqlQuery(sql)
+
+            return res.status(201).json({
+                status : "Created",
+                message : "berhasil menambahkan kategori pengaduan",
+                errors : [],
+                data : []
+            })
+
+
+        }catch(err){
+            return res.status(500).json({
+                status : "Internal Server Error",
+                message : "terjadi kesalahan diserver",
+                errors : [err.message],
+                data : []
+            })
+        }   
+    }
+
+    static async del(req,res){
+        try{
+
+            const {result}  = await mysqlQuery(`SELECT * FROM kategori_pengaduan WHERE id = ${req.params.id}`)
+
+            if(!result.length){
+                return res.status(404).json({
+                    status : "Not Found",
+                    message : "terjadi kesalahan diclient",
+                    errors : ["kategori pengaduan tidak ditemukan"],
+                    data : []
+                })
+            }
+
+            const img = result[0].foto.split("/")
+            const imgName = img[img.length - 1]
+
+            await mysqlQuery(`DELETE FROM kategori_pengaduan WHERE id = ${req.params.id}`)
+
+            const dirName = getDirName()
+
+            fs.unlinkSync(`${dirName}/images/${imgName}`)
+
+            return res.status(200).json({
+                status : "OK",
+                message : "berhasil menghapus pengaduan dan relasinya",
                 errors : [],
                 data : []
             })
